learning_assistant/requirements.txt
crewai>=0.28.8
python-dotenv>=1.0.0
openai>=1.0.0
